[Merge] Merge from develop to master

See merge request s11-webmobile1-sub2/S11P12A503!125

+++ backend/src/main/java/com/ssafy/picple/domain/calendar/service/CalendarServiceImpl.java
@@ -1,15 +1,25 @@
 package com.ssafy.picple.domain.calendar.service;
 
-import static com.ssafy.picple.config.baseResponse.BaseResponseStatus.*;
+import static com.ssafy.picple.config.baseresponse.BaseResponseStatus.*;
 
+import java.io.File;
+import java.io.FileNotFoundException;
+import java.io.IOException;
+import java.io.InputStream;
+import java.net.URL;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.time.LocalDate;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.ssafy.picple.config.baseResponse.BaseException;
+import com.ssafy.picple.config.baseresponse.BaseException;
 import com.ssafy.picple.domain.board.entity.Board;
 import com.ssafy.picple.domain.board.repository.BoardRepository;
 import com.ssafy.picple.domain.board.service.BoardService;
@@ -37,28 +47,57 @@ public class CalendarServiceImpl implements CalendarService {
 
 	// 캘린더 날짜(년월일)별 사진 개수 조회
 	@Override
-	public Long getPhotoCounts(Long userId, LocalDate createdAt) {
-		return calendarRepository.countByUserIdAndDate(userId, createdAt);
+	public Long getPhotoCounts(Long userId, LocalDate createdAt) throws BaseException {
+		try {
+			return calendarRepository.countByUserIdAndDate(userId, createdAt);
+		} catch (Exception e) {
+			throw new BaseException(DATABASE_ERROR);
+		}
+	}
+
+	// 캘린더 달력(년월)별 사진 개수 조회
+	@Override
+	public List<Long> getMonthlyPhotoCounts(Long userId, LocalDate monthlyStartDate, LocalDate monthlyEndDate) throws
+			BaseException {
+		try {
+			// 월별 날짜 리스트 생성
+			// datesUntil 메서드로 monthlyStartDate부터 monthlyEndDate-1 까지의 Stream 생성 -> 이후 리스트로
+			List<LocalDate> monthlyDates = monthlyStartDate.datesUntil(monthlyEndDate.plusDays(1)).toList();
+
+			// 사진 개수 저장할 리스트
+			List<Long> monthlyPhotoCounts = new ArrayList<>();
+			for (LocalDate date : monthlyDates) {
+				Long count = calendarRepository.countByUserIdAndDate(userId, date);
+				monthlyPhotoCounts.add(count);
+			}
+			return monthlyPhotoCounts;
+		} catch (Exception e) {
+			throw new BaseException(DATABASE_ERROR);
+		}
 	}
 
 	// 캘린더 일별 정보 조회
 	@Override
-	public List<CalendarDto> getDailyCalendars(Long userId, LocalDate createdAt) {
-		List<Calendar> calendars = calendarRepository.findByUserIdAndCreatedAt(userId, createdAt);
-		return calendars.stream()
-				.map(calendar -> {
-					Long photoId = calendar.getPhoto().getId();
-					// content를 PhotoUser에서 가져와야함 Calendar - Photo - PhotoUser로 연결되어있음
-					// 여기서 content가져와야함
-					PhotoUser photoUser = photoUserRepository.findByPhotoIdAndUserId(photoId, userId);
-					String content = photoUser != null ? photoUser.getContent() : "";
+	public List<CalendarDto> getDailyCalendars(Long userId, LocalDate createdAt) throws BaseException {
+		try {
+			List<Calendar> calendars = calendarRepository.findByUserIdAndCreatedAt(userId, createdAt);
+			return calendars.stream()
+					.map(calendar -> {
+						Long photoId = calendar.getPhoto().getId();
+						// content를 PhotoUser에서 가져와야 함 Calendar - Photo - PhotoUser로 연결되어 있음
+						// 여기서 content 가져와야 함
+						PhotoUser photoUser = photoUserRepository.findByPhotoIdAndUserId(photoId, userId);
+						String content = photoUser != null ? photoUser.getContent() : "";
 
-					return new CalendarDto(
-							calendar.getId(),
-							calendar.getPhoto().getPhotoUrl(),
-							content
-					);
-				}).collect(Collectors.toList());
+						return new CalendarDto(
+								calendar.getId(),
+								calendar.getPhoto().getPhotoUrl(),
+								content
+						);
+					}).collect(Collectors.toList());
+		} catch (Exception e) {
+			throw new BaseException(DATABASE_ERROR);
+		}
 	}
 
 	// 캘린더 선택 사진별 설명 작성
@@ -73,7 +112,7 @@ public class CalendarServiceImpl implements CalendarService {
 		Photo photo = calendar.getPhoto();
 
 		// 현재 접속자와 캘린더 작성자가 같은지 확인
-		if (userId == user.getId()) {
+		if (Objects.equals(userId, user.getId())) {
 
 			// 특정 사용자의 특정 사진 content찾기
 			PhotoUser photoUser = photoUserRepository.findByPhotoIdAndUserId(photo.getId(), userId);
@@ -81,6 +120,7 @@ public class CalendarServiceImpl implements CalendarService {
 			if (photoUser != null) {
 				photoUser.setContent(content);
 			} else {
+				// photoUser가 null
 				throw new BaseException(GET_PHOTO_USER_EMPTY);
 			}
 
@@ -102,7 +142,7 @@ public class CalendarServiceImpl implements CalendarService {
 		Photo photo = calendar.getPhoto();
 		User user = calendar.getUser();
 
-		if (userId != user.getId()) {
+		if (!Objects.equals(userId, user.getId())) {
 			throw new BaseException(NOT_EQUAL_USER_ID);
 		}
 
@@ -127,6 +167,52 @@ public class CalendarServiceImpl implements CalendarService {
 		}
 	}
 
+	// 캘린더에서 특정 사진 선택하여 로컬에 다운로드
+	@Override
+	public void downloadPhoto(Long calendarId, Long userId) throws BaseException {
+		Calendar calendar = calendarRepository.findById(calendarId)
+				.orElseThrow(() -> new BaseException(GET_CALENDAR_EMPTY));
+
+		User user = calendar.getUser();
+		if (userId.equals(user.getId())) {
+			// 다운로드 받을 사진 경로와 저장될 default 이름 설정(중복고려)
+			String photoUrl = calendar.getPhoto().getPhotoUrl();
+			String fileName = "picple-download" + "_" + System.currentTimeMillis() + ".jpg";
+
+			/**
+			 * 여러 운영체제 호환 고려한 다운로드 경로 -> home디렉토리 가져오고 그 뒤에 폴더 추가
+			 * openStream으로 URL에 대한 입력 스트림을 열고 URL이 가리키는 자원 데이터를 읽기
+			 * copy메서드로 경로에서 파일 읽어와 특정경로에 복사
+			 */
+			String downloadDirPath =
+					System.getProperty("user.home") + File.separator + "Downloads" + File.separator + fileName;
+			Path photoPath = Paths.get(downloadDirPath);
+
+			try (InputStream in = new URL(photoUrl).openStream()) {
+				Files.copy(in, photoPath);
+			}  catch (FileNotFoundException e) {
+				throw new BaseException(FILE_NOT_FOUND_ERROR);
+
+			} catch (IOException e) {
+				e.printStackTrace(); // 디버깅을 위한 로그
+
+				throw new BaseException(FILE_DOWNLOAD_ERROR);
+			}
+
+		} else {
+			throw new BaseException(NOT_EQUAL_USER_ID);
+		}
+
+	}
+
+	// 캘린더Id에 해당하는 PhotoUrl 조회(사진 다운로드용)
+	@Override
+	public String getPhotoUrlByCalendarId(Long calendarId, Long userId) throws BaseException {
+		Calendar calendar = calendarRepository.findById(calendarId)
+				.orElseThrow(() -> new BaseException(GET_CALENDAR_EMPTY));
+		return calendar.getPhoto().getPhotoUrl();
+	}
+
 	// 캘린더에서 사진 삭제
 	// 삭제시 board에서도 삭제됨.
 	@Override
@@ -140,6 +226,10 @@ public class CalendarServiceImpl implements CalendarService {
 		Board board = boardRepository.findByUserIdAndPhotoIdAndIsDeletedFalse(calendar.getUser().getId(),
 				calendar.getPhoto().getId());
 
-		boardService.deleteBoard(board.getId(), board.getUser().getId());
+		// board에도 존재하면 여기서도 삭제 진행
+		if (board != null) {
+			boardService.deleteBoard(board.getId(), board.getUser().getId());
+		}
+
 	}
 }
